Main

Description

Predicts the next word based on an input and a given model.

NGramPredict

main :: IO ()

Gets flags and validates them. Required flags are ´NGramPredict <number> <model> <file> <line> <column>´

IO

validate

Arguments

:: [String]
runtime flags

-> IO ()
return value

Controls whether or not enoughs arguments are given, the base functions themselves test if the arguments are of the right type

getNGram

Arguments

:: String
path to text file

-> Int
line

-> Int
column

-> IO [String]
return value

A 4-gram is build based on the given line and column. Should the function encounter certain punctuation, a shorter n-gram is build. No n-gram is build when the given word is followed up by punctuation.

getList

Arguments

:: Int
length of n-grams to search for

-> String
model file path

-> IO [[String]]
return value

Given a valid model, the function builds a list with all n-grams of a certain length.

write

Arguments

:: Int
number of values to print

-> [[String]]
list of n-grams

-> IO ()
return value

Writes a certain number of predictions from a list of n-grams.

Predict

loop

Arguments

:: [String]
runtime flags

-> [String]
n-gram

-> [[String]]
list of predicted n-ngrams

-> IO ()
return value

Loops through all possible sizes of n-grams and builds a list of predictions. Results from longer n-grams appear first in the list. If more predictions are used than actually found, the most common 1-grams are searched.

filter

Arguments

:: [String]
(n-1)-gram

-> [[String]]
list of n-ngrams

-> [[String]]
return value

Takes a list of n-grams and one (n-1)-gram. The resultig list contains all n-grams the is (n-1)-gram is a prefix of.

sort

Arguments

:: Int
number of elements to sort

-> [[String]]
list of n-grams

-> [[String]]
return value

Performs a linear search until the sorted list is exactly the asked length. All remaining unsorted elements are dropped.