Main

Description

Predicts the next word based on an input and a given model.

NGramPredict

main :: IO ()

Gets flags and validates them. Required flags are ´NGramPredict <number> <model> <file> <line> <column>´

IO

validate

Arguments

:: [String]
runtime flags

-> IO ()
final results

Controls whether or not enoughs arguments are given, the base functions themselves test if the arguments are of the right type.

write

Arguments

:: [[String]]
list of ["prob", "prediction", "bow"]

-> IO ()
final console output

Writes predictions from a list.

Prediction

getNGram

Arguments

:: String
input file

-> Int
line

-> Int
column

-> Int
max length

-> [String]
final n-gram

Builds a n-gram based on an input file and a position. Its length is based on the maximum length of n-grams in the model. When encountering punctuation, the n-gram stops regardless of its length.

getModel

Arguments

:: String
content or .arpa file

-> Map [String] [[String]]
map of .arpa file

Converts a .arpa file into a map of the form [elements of (n-1)-gram] [["probability", "n-th element", "bow"]], i.e. one key holds all possible predicions with their probabilites as value.

predict

Arguments

:: Map [String] [[String]]
map of .arpa file

-> [String]
current prefix

-> [String]
current n-gram

-> [[String]]
list [["prob", "pred", "bow"]]

Gets all values using the current n-gram as key, then drops the first element and searches again, correcting the probability with the bow off all prior dropped elements.